services/agents/requirements.txt
# AI Startup Co-Founder Agents Requirements

# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# LLM and AI
google-generativeai==0.8.3
openai==1.6.1
langchain==0.0.350
langchain-openai==0.0.2
transformers==4.36.0
torch==2.6.0

# HTTP and API
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Database and storage
asyncpg==0.29.0
redis==5.0.1
aiofiles==23.2.1

# Vector databases
faiss-cpu==1.12.0
pinecone-client==2.2.4
weaviate-client==3.25.3

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Environment and configuration
python-dotenv==1.0.0
click==8.1.7

# Data processing
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.2

# File handling
python-magic==0.4.27

# CORS and middleware
fastapi-cors==0.0.6
